Move the flow analysis from the Utilites to its own folder
